Fix/user photo

+++ src/services/users/userCreate.services.ts
@@ -3,11 +3,8 @@ import { hash } from "bcryptjs";
 import { AppError } from "../../errors/AppError";
 import { IUserRequest } from "../../interfaces/users";
 import { User } from "../../entities/users.entity";
-import { v4 as uuid } from "uuid";
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
-import path from "path";
-import handleErrorMiddleware from "../../middlewares/handleError.middleware";
 
 const userCreateService = async ({
   username,
@@ -25,25 +22,32 @@ const userCreateService = async ({
     throw new AppError(409, "This email already exists");
   }
 
-  const cloudinaryFile: any = await cloudinary.uploader.upload(
-    photo.path,
-    (error: any, result: any) => {
-      if (error) {
-        throw new AppError(500, `Internal server error, ${error}`);
+  if (photo) {
+    const cloudinaryFile: any = await cloudinary.uploader.upload(
+      photo.path,
+      (error: any, result: any) => {
+        if (error) {
+          throw new AppError(500, `Internal server error, ${error}`);
+        }
+        return result;
       }
-      result;
-    }
-  );
+    );
 
-  fs.unlink(photo.path, (error) => {
-    if (error) {
-      console.log("erro");
-      throw new AppError(500, `Internal server error ${error}`);
-    }
-  });
-  photo = cloudinaryFile.public_id;
+    fs.unlink(photo.path, (error) => {
+      if (error) {
+        console.log("erro");
+        throw new AppError(500, `Internal server error ${error}`);
+      }
+    });
+    photo = cloudinaryFile.public_id;
+  }
+
+  if (!photo) {
+    photo = "Imagem padrão";
+  }
 
   const hashedPassword = await hash(password, 10);
+
   const user = userRepository.create({
     username,
     email,
